+++ internal/provider/sdkv2/service_packages_gen.go
@@ -12,6 +12,7 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/service/accessanalyzer"
 	"github.com/hashicorp/terraform-provider-aws/internal/service/account"
+	"github.com/hashicorp/terraform-provider-aws/internal/service/accountaccess"
 	"github.com/hashicorp/terraform-provider-aws/internal/service/acm"
 	"github.com/hashicorp/terraform-provider-aws/internal/service/acmpca"
 	"github.com/hashicorp/terraform-provider-aws/internal/service/agentregistry"
@@ -288,6 +289,7 @@ func servicePackages(ctx context.Context) []conns.ServicePackage {
 	v := []conns.ServicePackage{
 		accessanalyzer.ServicePackage(ctx),
 		account.ServicePackage(ctx),
+		accountaccess.ServicePackage(ctx),
 		acm.ServicePackage(ctx),
 		acmpca.ServicePackage(ctx),
 		agentregistry.ServicePackage(ctx),