+++ internal/service/resiliencehubv2/sweep.go
@@ -28,7 +28,10 @@ func RegisterSweepers() {
 		"aws_resiliencehubv2_service_function",
 	)
 	awsv2.Register("aws_resiliencehubv2_service_function", sweepServiceFunctions)
-	awsv2.Register("aws_resiliencehubv2_system", sweepSystems)
+	awsv2.Register("aws_resiliencehubv2_system", sweepSystems,
+		"aws_resiliencehubv2_user_journey",
+	)
+	awsv2.Register("aws_resiliencehubv2_user_journey", sweepUserJourneys)
 }
 
 func sweepPolicies(ctx context.Context, client *conns.AWSClient) ([]sweep.Sweepable, error) {
@@ -202,6 +205,44 @@ func sweepAssertions(ctx context.Context, client *conns.AWSClient) ([]sweep.Swee
 					sweepResources = append(sweepResources, framework.NewSweepResource(newAssertionResource, client,
 						framework.NewAttribute("service_arn", serviceARN),
 						framework.NewAttribute("assertion_id", aws.ToString(v.AssertionId)),
+					))
+				}
+			}
+		}
+	}
+
+	return sweepResources, nil
+}
+
+func sweepUserJourneys(ctx context.Context, client *conns.AWSClient) ([]sweep.Sweepable, error) {
+	conn := client.ResilienceHubV2Client(ctx)
+	var input resiliencehubv2.ListSystemsInput
+	var sweepResources []sweep.Sweepable
+
+	pages := resiliencehubv2.NewListSystemsPaginator(conn, &input)
+	for pages.HasMorePages() {
+		page, err := pages.NextPage(ctx)
+		if err != nil {
+			return nil, smarterr.NewError(err)
+		}
+
+		for _, v := range page.SystemSummaries {
+			systemARN := aws.ToString(v.SystemArn)
+			input := resiliencehubv2.ListUserJourneysInput{
+				SystemArn: aws.String(systemARN),
+			}
+
+			pages := resiliencehubv2.NewListUserJourneysPaginator(conn, &input)
+			for pages.HasMorePages() {
+				page, err := pages.NextPage(ctx)
+				if err != nil {
+					return nil, smarterr.NewError(err)
+				}
+
+				for _, v := range page.UserJourneySummaries {
+					sweepResources = append(sweepResources, framework.NewSweepResource(newUserJourneyResource, client,
+						framework.NewAttribute("system_arn", systemARN),
+						framework.NewAttribute("user_journey_id", aws.ToString(v.UserJourneyId)),
 					))
 				}
 			}